Logging: Move logging to own module
Moves the `configure_logging` function from `misc/constants`
to `output/Logging`.
