+++ coalib/coala_delete_orig.py
@@ -2,12 +2,12 @@ import os
 
 from pyprint.ConsolePrinter import ConsolePrinter
 
+from coalib.output.Logging import configure_logging
 from coalib.output.printers.LogPrinter import LogPrinter
 from coalib.parsing import Globbing
 from coalib.settings.ConfigurationGathering import get_config_directory
 from coalib.settings.Section import Section
 from coalib.parsing.Globbing import glob_escape
-from coalib.misc.Constants import configure_logging
 
 
 def main(log_printer=None, section: Section=None):
