+++ coalib/misc/Constants.py
@@ -128,39 +128,3 @@ URL_REGEX = re.compile(
     r'(?::\d+)?'  # optional port number
     r'(?:/?|[/?]\S+)$',  # path
     re.IGNORECASE)
-
-
-def configure_logging():
-    """
-    Configures the logging with hard coded dictionary.
-    """
-    import sys
-    import logging.config
-    logging.config.dictConfig({
-        'version': 1,
-        'handlers': {
-            'colored': {
-                'class': 'logging.StreamHandler',
-                'formatter': 'colored',
-                'stream': sys.stderr
-            }
-        },
-        'root': {
-            'level': 'DEBUG',
-            'handlers': ['colored']
-        },
-        'formatters': {
-            'colored': {
-                '()': 'colorlog.ColoredFormatter',
-                'format': '%(log_color)s[%(levelname)s]%(reset)s[%(asctime)s]'
-                          ' %(message)s',
-                'datefmt': '%X',
-                'log_colors': {
-                    'ERROR': 'red',
-                    'WARNING': 'yellow',
-                    'INFO': 'blue',
-                    'DEBUG': 'green'
-                }
-            }
-        }
-    })
